refactor: feature의 model 폴더명 models로 변경
