+++ frontend/src/pages/Join/JoinStep3Password.tsx
@@ -6,7 +6,7 @@ import Layout from '@/layout/Layout';
 import { JoinTemplate, IUserRegistration, useAuth } from '@/features/auth';
 import { AlertText, Input } from '@/shared/components';
 import { withUnauthenticatedUser } from '@/shared/hocs';
-import { VALIDATE } from '@/constants/validate';
+import { VALIDATE } from '@/shared/constants/validate';
 
 const JoinStep3Password = () => {
   const { errorMessage, userJoin } = useAuth();
